add configuration
