+++ config.yaml.example
@@ -1,0 +1,3 @@
+urlhandlers:
+  AlbumHandler:
+    base_path: "~/Music/"
