+++ reader.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
 SCRIPTDIR=$(dirname "$0")
-if [[ ! -d $SCRIPTDIR/venv ]]; then
+pushd $SCRIPTDIR
+
+if [[ ! -d venv ]]; then
   >&2 echo run \'python3 -m venv venv\' in $SCRIPTDIR directory before calling $0
   exit 1
 fi
 
-source $SCRIPTDIR/venv/bin/activate
-$SCRIPTDIR/reader.py
+venv/bin/python reader.py
